Area・Genreのコントローラーに登録画面・登録機能を追加。（AdminControllerはそのコードを削除）
