+++ app/Http/Controllers/AreaController.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+use App\Models\Area;
+
+
+class AreaController extends Controller
+{
+    public function areaView()
+    {
+        return view('area_register');
+    }
+    public function areaRegister(Request $request)
+    {
+        $area = Area::create([
+            'name' => $request->name,
+        ]);
+        unset($area['_token']);
+        return redirect('/admin');
+    }
+}
